Store responses as ints of the TXT data; use Ticker
Changed the comparrison to extract the value of the TXT record so we
ignore the TTL, dropped the use of reflect.DeepEqual, did a simple ">=
master" check.  This version is broken because the pointer to the int
isn't shared correctly.  I'll refactor that out to a single goroutine in
the next commit, but wanted to get this chunk checked in as one piece.